Removed conditional statements from puppet/passenger.pp
Originally the resources in puppet/passenger.pp would not be applied
if catalog_cert_autosign was set to false. This was set to prevent
httpd from failing to start when an autosign entry was not created for a
Catalog Master. This created a requirement on catalog_cert_autosign being
set to true if you wished to ever bring up your Catalog Master. I now
rather have httpd fail to start, than the requirement of a parameter
being true (confusing!).