Removed unnecessary code from puppet/role/catalog.pp
Removed a notify message about having catalog_cert_autosign set to false,
as it's again forcing a parameter to be true. Also changed an inline command
being used in an unless, to a static path. There's no point in trying to
have the check dynamic, when the provisions aren't there to allow people
to easily change the default $vardir.