Removed ability to set environments directory
Removed the ability to specify a different environments directory as there isn't a good
reason at this time to allow it. It just added more overhead, and another parameter.
I also modified the default value for the conf_master hash to a simpler configuration.
This will force anyone that wants to use directory environments, to override the
appropriate conf_* parameter hashes.