Added more validations & pm_server parameter
Realized that if the puppet_role was set to 'ca', I wasn't forcing a server
option to be set in puppet.conf. Similar to what I was doing with ca_server
for the catalog role. More validations were also added to ensure that a
custom conf_(main,agent) hash contains the appropiate keys for the specified
puppet_role. The default values for the conf_* hashes have also been updated
appropriately.