Implemented a new class parameter, puppet_vardir
Instead of echo'ing the home of the current user to set :puppetdir in
/etc/puppet/node.rb (dumb), I have created a new class parameter that will
be used to set $vardir everywhere. It's also being used to set the home of
the Puppet user. Added a file resource for $puppet_vardir to ensure it
exists as other file resources down the line rely on it.