Modified acceptance tests
Modified the default manifest after the default values were changed for
the foreman and smartproxy parameters. Added a sed statement that will
point puppet-stack.local to the SUT. This allows `puppet agent -t`
to run successfully if users try it by bringing up the SUT with
BEAKER_destroy=no and SSH'ing in.