Fixed a case where $foreman_ssl_ca was not being set properly
If this module is just used to configure the Foreman, the $foreman_ssl_ca
variable in foreman.pp needs to use the ca cert in /etc/puppet/ssl, as it's
assumed that the Puppet node is a client and not a master. This is a slight
fix to the changes made in hotfix-0.0.5.