Updated default branches, added 2 new parameters
The default branch for the Foreman & smart-proxy have both been changed
to 1.6-stable. The Foreman 1.6+ now generates a default admin password
during the rake db:seed task, and spits it to stdout. Since this is not easily
visible, we now declare the default password using foreman_default_password.
Also added foreman_try_rake_tasks so that it's possible to control if
the Foreman's rake tasks will run.